+++ src/index.ts
@@ -3,10 +3,14 @@ import * as ts_module from 'typescript/lib/tsserverlibrary';
 import { Logger } from './logger';
 import { ConfigurationManager } from './settings';
 import * as mockRequire from 'mock-require';
+import { WorkspaceLibraryExecution } from './runner';
+
+const enableWorkspaceLibraryExecutionEnvVar = 'TS_TSLINT_ENABLE_WORKSPACE_LIBRARY_EXECUTION';
 
 export = function init({ typescript }: { typescript: typeof ts_module }) {
     const configManager = new ConfigurationManager(typescript);
     let logger: Logger | undefined;
+    let plugin: TSLintPlugin | undefined;
 
     // Make sure TS Lint imports the correct version of TS
     mockRequire('typescript', typescript);
@@ -24,13 +28,26 @@ export = function init({ typescript }: { typescript: typeof ts_module }) {
                 return info.languageService;
             }
 
-            return new TSLintPlugin(typescript, info.languageServiceHost, logger, info.project, configManager)
-                .decorate(info.languageService);
+            plugin = new TSLintPlugin(typescript, info.languageServiceHost, logger, info.project, configManager);
+
+            // Allow clients that don't use onConfigurationChanged to still securely enable
+            // workspace library execution with an env var.
+            const workspaceLibraryFromEnv = process.env[enableWorkspaceLibraryExecutionEnvVar] ? WorkspaceLibraryExecution.Allow : WorkspaceLibraryExecution.Unknown;
+            plugin.updateWorkspaceTrust(workspaceLibraryFromEnv);
+
+            return plugin.decorate(info.languageService);
         },
         onConfigurationChanged(config: any) {
             if (logger) {
                 logger.info('onConfigurationChanged');
             }
+
+            if (plugin) {
+                if ('allowWorkspaceLibraryExecution' in config) {
+                    plugin.updateWorkspaceTrust(config.allowWorkspaceLibraryExecution ? WorkspaceLibraryExecution.Allow : WorkspaceLibraryExecution.Disallow);
+                }
+            }
+
             configManager.updateFromPluginConfig(config);
         },
     };