+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.0.0 - November 30, 2020
+- Restricts when tslint is loaded from the workspace.
+    - Global TSLint versions can always be loaded.
+    - TSLint versions installed next to the plugin can always be loaded.
+    - Otherwise, the consumer of the plugin must use `onConfigurationChanged` and explicitly enable `allowWorkspaceLibraryExecution`.
+    - You can also force allow workspace versions of TSLint to be loaded by setting a `TS_TSLINT_ENABLE_WORKSPACE_LIBRARY_EXECUTION`  environment variable.
+
 ## 0.5.5 - November 11, 2019
 - Restore old cwd after linting finishes.